Move `ParameterGenerationError` to a separate file
