+++ android/src/main/kotlin/net/mullvad/mullvadvpn/model/ParameterGenerationError.kt
@@ -1,0 +1,8 @@
+package net.mullvad.mullvadvpn.model
+
+sealed class ParameterGenerationError {
+    class NoMatchingRelay : ParameterGenerationError()
+    class NoMatchingBridgeRelay : ParameterGenerationError()
+    class NoWireguardKey : ParameterGenerationError()
+    class CustomTunnelHostResultionError : ParameterGenerationError()
+}
